pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-ops-dashboard"
version = "0.1.0"
description = "Personal Ops Dashboard API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<0.116",
  "uvicorn[standard]>=0.32,<0.33",
  "pydantic-settings>=2.6,<2.7",
  "sqlalchemy>=2.0,<2.1",
  "alembic>=1.14,<1.15",
  "psycopg[binary]>=3.2,<3.3",
  "python-jose[cryptography]>=3.3,<3.4",
  "bcrypt>=4.0",
  "python-multipart>=0.0.17,<0.0.18",
  "email-validator>=2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<8.4",
  "anyio[trio]>=4.0",
  "pytest-cov>=6.0,<6.1",
  "httpx>=0.28,<0.29",
  "ruff>=0.7,<0.8",
  "black>=24.10,<24.11",
  "mypy>=1.13,<1.14",
  "pre-commit>=4.0,<4.1",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
